Update 016. 3Sum Closet
Added two "checks" to check if we can stop the inner while loop earlier. That is, for example, [1, 100, 200, 300, 400, 500] and target is 1000000. When you iterate: the value at "i" is 1, "start" is 200 and "end" is 500, so obviously there is no use iterating from "start" to "end", because even for (1 + 400 + 500) = 901, which is still far smaller than the target. 

In the case above, you just update the error using "901" and break the loop.